Add support for debugging from CLI
This diff adds a delve debugger, and disables hardening, which is
needed because delve builds C code with -O0 and hardening conflicts
with it and requires -O2.

You can run the debugger like this:

```
cd tests
dlv test
```

And you'll be presented with a CLI and able to set a breakpoint.