Get rid of transaction conflicts between shards
This diff implements a quick and dirty workaround for transaction
conflicts between shards. By "quick and dirty" I mean not in terms
of the overall approach, but more or less in terms of code that may
not be all that beautiful.

It solves the problem though:

- First, it separates everything except blocks into their own namespace
- Second, it passes shardId to ExecutionState

This means that whatever the transactions write can essentially
never clash.