Add static linking and packaging
This diff makes a portable `.deb` package in the same sense as we had
for DBMS. The files in the package are platform-independent and
  statically linked.

If you wonder why I had to include `musl` in the diff, the reason is
actually straightforward: we use some "CGO" extensions (at least SQLite
at the moment). And those extensions are compiled with the C/C++
toolchain and linked. By default that means that we depend on a bunch of
system packages, and if you do `ldd` on the resulting binary, you'd
clearly see that. (This is even though Go _itself_ doesn't link against
external things)

This diff doesn't require you to change any habits about your
development, and concerns the CI and deployment only.