update message struct and import crypto stuff for it
This patch removes several fields from message - e.g. ShardId since it should be
possible to extract it from address. Then we need to remove message index since it's
impossible to specify it on the user side.
Also we need GasPrice and GasLimit fields in our economy.

Then this patch imports some crypto stuff to be able to sign and verify messages.