move "internal" message field to beginning
This patch implements a request to move "Signature" field to the end of the message
structure. It's reasonable because during encoding it's possible to prepare digest,
sign it and only then add Signature to blob.

So the single reasonable plase for such flag is the beginning of the message.
Because it's possible to determine its type during decoding (it will be useful
if we will have different structure for internal and external messages).

Also this patch fixes SSZ generation. The only significant change that was applied
is moving of AddrSize constant to "types" package. Otherwise fastssz considered
it as dynamic sized field (with offsets and other undesirable effects).