Add an example of deploying and calling a simple contract via CLI
Recently, a CLI tool for deploying and calling smart contracts to the
node has been merged to the master branch.

This PR adds a step-by-step example to the README.md on how to compile
a simple counter smart contract, deploy it and call it. The contract
uses logs to communicate the current counter value.

Note that as of now, you need to specify the method ID manually,
because there is no ABI decoding in the CLI yet. But we can work on
that to make interaction more obvious.