Specify in README that the determinate installer for Nix should be used
Currently, the README.md says that the Determinate Systems installer
for Nix should be only used on Apple Silicon, which is wrong. The
official installer doesn't enable flakes by default, which means that
you need to do extra tuning of configuration to make things in this
guide work.

The Determinate Systems installer works on all platforms, so I don't
see any reason to use anything else.