split storage types and RPC types
This patch removes usage of basic types from RPC. The main issue here that
we stored a lot of redundant fields in Receipt and Log structures.

Also after this patch we will also have strict types for our JSON RPC
that is also quite useful for our nil.js toolchain.

Closes #142