Compile static binaries on all platforms
We used to use CGO because we were linking with SQLite. But at the
moment we don't depend on any external libraries (only on pure Go).

Because of this, if we disable CGO and switch `-linkmode` to be
"internal", we get static binaries on all platforms without having to
use static toolchain (which doesn't work on Mac).