Switch to using vendored go.mod in "nix develop"
This diff was extracted from another diff I've been working
on (explorer).

I noticed that there's a difference between how we use dependencies in
"nix develop" vs "nix build". In "nix build", dependencies are taken
from the "vendored" directory. This means that dependencies are not
downloaded, but taken from a local directory. This is expected, and
generally how Go builds work in Nix.

For "nix develop" however, we were setting a number of env variables
by hand and used dynamic downloads. This sometimes leads to problems
where what you use in "nix develop" diverging from "nix build".

One such example is "sszgen". This tool was loaded fine by "nix
develop", because Go would dynamically determine that it needs to
download the module. But if you try to use it in "nix build", it will
fail due to a vendoring error.

To understand why this is the case in Go, you can read this article:
https://www.jvt.me/posts/2022/06/15/go-tools-dependency-management/

TL;DR is that you should put tool dependencies into tools.go. If you
do this, they become part of the vendoring process as expected, and we
can have the same approach for "nix build" and "nix develop"