Fix test execution in "nix develop"
There are two fixes in this diff.

First is that we used `runtime.Caller()` to calculate the base
directories to locate files from tests. This is a wrong approach,
because what `runtime.Caller()` returns depends on how you build your
code. It turns out that the file name it returns may contain the
repository prefix (which for obvious reasons doesn't exist on the
filesystem).

Thankfully, by convention, go tests set the location of the test file
as `CWD`, meaning that we can rely on that to be the case. And so we
don't need any tricks. Just calculate the path relative to CWD and
take the absolute path after that.

The second issue is more tricky. I had tests failing with

```
signature_nocgo.go:83: test panicked: private key curve is not secp256k1
```

This is because of this line:
https://github.com/ethereum/go-ethereum/blob/master/crypto/signature_nocgo.go#L83

We initialized private key's Curve object by using a different
library, while the `signature_nocgo.go` compares the Curve with
`S256()` object that it has. In case when we build without CGO, those
objects are not equal, even though logically this is still correct,
and we are using sec256k1. To fix this, I'm using the S256 object from
go-thereum/crypto.