tests: basic infrastructure for cli integration tests
This patch adds minimal infrastructure necessary for integration testing the CLI.

Previously, we only tested cli/service, which is better than nothing but does
not catch errors at the highest level of interaction with Cobra. Therefore,
tests that directly invoke the CLI are needed.

The first option was to move all contents of cmd/nil_cli to a separate module
and run Cobra manually. However, this approach had a major drawback: all data
is written to a logger that writes to stdout, which cannot be easily verified.

Thus, the second option was chosen - simply call go run and execute nil_cli in
a separate process. This approach takes longer but is the most accurate from a
testing perspective.