Fix "nil_cli wallet send-message" default amount being non-zero
In the CLI parameter parsing, there is an error when `nil_cli wallet new`
and `nil_cli wallet send-message` reuse the same parameter variable
for specifying the amount of tokens to send.

In `nil_cli wallet send-message, the default value is 0, which it
should be. But it gets overriden by the other command.

What this does is that if you don't specify `--amount` and call a
non-payable method, you'll essentially still pass about 10m tokens in
the message, and there would be "execution reverted" error down the
line.

This patch is simple -- let's use different variables for the two
different commands.