Fix a bug of fetching receipts outmessages if multiple messages per block
There was an arithmetic error in the JsonRPC where in case we had
multiple messages handled per block, we were sometimes not outputting
outmessages depending on the position of the message in the block.

receipt.OutMsgIndex shows the start index of the outmessage, while
receipt.OutMsgNum shows the number of messages in receipt (and not the
end index as the code supposed.

We've found the issue on prod today due to the fact that there's load
generator producing messages in addition to the user. Found by Aleksa.

We didn't see the error locally for obvious reasons (people only
submit transactions one-by-one.