Better CLI output (#504)
The idea of our outputs is the following:

By default, logs are disabled. They can be enabled with the verbose flag
We provide the quiet flag that disables messages with values explanation, leaving only bear values, so that output can be used in pipes/scripts/etc.
Errors are always written to stderr. They are also logged in some places where it would make them easier for debugging
So by default the user doesn't see the logs, and they actually are not meant to be understandable by the user. They can be used by us, when the user files an issue, he should provide verbose logs to us.
With verbose logs we can see the same info printed twice in logs and in the stdout, but it's ok, because one is service logs and the other is command output