Add manual timestamp managing support to the db (#574)
Here, I make our Badger to be able to work with timestamps. For that, I forked Badger and allowed the creation of read-only transactions with a provided timestamp in the non-managed mod. Also, I added a way to retrieve the commit timestamp. With these two things, we will be able to make a mirroring work properly.