Read-through database support (#563)
This PR adds read-through database support
./nil --read-through-db-addr <addr> --read-through-start-block <block>
The proxy database reads the provided block from the MainChain and then initializes the last blocks of every chain according to the content of ChildBlocks. Then, it just proxies the requests using the right timestamp. For now, it doesn't cache anything when proxying; I will add this as a follow-up. The semantics should be apparent from the test.