feat: add proof creation/verification to mpt
Add methods to mpt package, related to Merkle proof creation and verification. Basicaly, Proof is just a slice of Nodes. This means, it is self-sufficient, and you could verify if without additional requests to db. Creation is basically a copy of mpt.get method, but appends traversed nodes to the slice.
Proof provides a method to construct in-memory Sparse MPT from it, setting the first slice element as a root. Later, it could be used to verify the existence of the key in the proof, returning its value.