feat: add GetContract method to dbg API
This methods is needed for zkEVM proof creation. We decided that it would be cheaper to get entire contract storage than make many requests related to each storage slot touched during the period being proved.
Contract storage is sent as a JSON array of key-value pairs, proof is sent as a JSON array of strings, where each string is SSZ-serialized Node, encoded to hex. Example of decoding could be seen in tests.