accept mainchain hash for eth_call
In future we will simulate cross-shard calls.
We need to specify a common reference point for all shards.
The only source of this information is the mainchain,
which contains hashes of blocks from other shards.
This patch changes the value of the blockNumOrHash argument.
Previously, it was a block used for calculations.
Now it is the main chain block.