cli: add abi helpers
In case if someone wants to analyze complex exectution flow it can be not enough
to have default decoding from `contract call-readonly` because it decodes just
top-level result. But outbound messages and their results can be represented as
raw binary data (because only contract author/user know ABI of nested calls).

So this patch introduces simple toolchain to be able to work with EVM ABI.
Just two functions - `encode` that accepts method name and its arguments
and `decode` that accepts method name and result of its execution.