Don't rebuild nil binaries unless necessary
This diff contains one main change, and one tangential fix.

The main change is to not rebuild nil binaries unless there was
actually a change somewhere in the source code. Currently if you're
working for example on docs and try to rebuild them, nild will be also
rebuilt as docs are currently part of the source input for nil.nix.

This is actually quite easy to fix with lib.sourceByRegex, which will
only keep what's necessary in the nil source tree. And as such it will
not be rebuilt unnecessarily.

Second is that I just noticed that we still load nil-hardhat plugin
from github repo and not from local source. Seems we've just forgot to
re-point `src` after absorbing the code into this repo. I've fixed this.