use api calls for hendling output messages
We will split different shard on different servers. So we will
have only main and some specific shard on single machine.
To avoid this issue we should proxy outbound messages to some
specific shards. But for now we can only rework API to call API
recursively (in future it will be network calls).