implement eth_estimateFee API
This patch adds the eth_estimateFee API to estimate funds
needed for sending a message. This call is not identical
to the original eth_estimateGas because gas prices may
vary across shards. Therefore, in our case, we get an
estimate of the funds required for the near future (and
this may easily change over time).

The algorithm is fairly simple, though it needs
explanation. We cannot simply execute eth_call with an
overridden contract balance because the gas consumed may
be less than what is required during the call (see EIP2200).
So, we override the account balance and use binary search
to vary feeCredit. A separate issue here is that this
process must be done for all messages sent. That is, if
we have an external message that leads to an internal
message, we must first estimate the fee for the internal
message and then for the external message. We do not
unpack messages and try not to touch user data. This can
be done later.

In tests, hardcoded constants were replaced with
empty/null values. If the user does not explicitly
specify a value, fee estimation is triggered. This allows
for more comprehensive code testing. However, it is not
possible to completely eliminate this, as we have
negative tests that check for errors. It is expected that
this will happen asynchronously, while fee estimation is
a fully synchronous operation.