Run tests in parallel in github actions
Up until recently, we were running all checks with `nix flake check -L`.
This meant that those checks were queued one after the other and spent
quite a bit of time (up to 20 minutes). Check are in fact independent
and can be run in parallel with a matrix job (which I ended up doing
in this pull request). As a bonus, you now know what exactly broke,
and don't need to hunt for the logs.