make rpc server optional
In future we won't need it at all. E.g. when node works as a
validator all interaction will be performed over binary protocol.

Also this patch changes option name to `--http-port`. For now
`--port` also works but it's only for backward compatibility.
The reason that we have libp2p `--tcp-port` and `--port` option
looks strange.