Fix `[]byte` arguments corruption in precompile contracts
When we do `args := abi.Unpack(input)`, some args with `[]byte` type (bytes in Solidity) are not
created as a separate memory array but just point to the slice inside the input calldata data.
Thus, when we save this data somewhere, it can be overwritten by executing the MSTORE instruction.