rename LocalRawApi to NodeRawApi and generalize it
This patch makes a bit clear why do we need to have "global" and
"shard" APIs as a separate entities.
Each node in any way will have "NodeRawApi" but under the hood this API
performs requests to local database (if we have local shard copy) or
to the network (via libp2p). It will be extremely useful when we will implement
eth_call/eth_estimateFee functions where a call can send one or several messages
that should be processed by remote instances.