add protobuf serialization for local calls
It's pretty hard to cover all (network and local paths) during tests in scope of single
instance. It would be useful to connect network peer to itself and call it but libp2p
prohibits it. So this patch adds new layer to the local part that behaves like
client + server. It serializes and deserializes call arguments and results to simulate
network call.
In that way we cover all parts - network and local at the same time.