add typed `messages` and `messageHashes` to `RPCBlock` type
This patch adds typing for messages in the block.
Previously, depending on the `fullTx` flag, you could
get a `[]any` array containing either hashes or messages.
After this patch, only messages will be in the array if
the `fullTx` flag is set. Without it, the `messageHashes`
field will be returned containing the message hashes.