add SendRawTransaction and GetTransactionCount calls to rawapi
This patch implements two essential calls for the Lite RPC node.
After applying this patch, it should be possible
to deploy contracts via the Lite RPC node.

The suggested solution is somewhat controversial because
the `msgpool` should probably be an independent entity. However,
for simplicity, we place it near the validator in this case,
just to unblock the normal usage of the Lite RPC node.