Fix deployment through sandbox to shards greater than one
There is a bug in the sandbox frontend where we send a deployment
command to the wallet, but wait for the receipt from the shard
that we deployed to. The issue is that the receipt should be
fetched from the shard where the wallet itself is located. Otherwise
we'll always get null and the deployment process will hand forever.

I've found this issue while trying to understand why I can't deploy
a smart contract through the sandbox to any shard other than the
first.