drop sleeps from tests
The main goal of this patch is to fix flaky tests. But actually tests worked fine but
we had port collision. And it was completely unclear in fact.
This patch splits allows to replace old approach with `go nilservice.Run()` with
```
node = CreateInstance()
node.Run()
```
It's pretty similar how subprocess.Popen works in Python.

In that case we will get an error significantly earlier before instance start.