feat: add MPT tracer (storage + contract trie) to synccommittee, Remove custom account state from tracer.
To be able to use execution.AccountState with tracer State DB, I had to slightly adjust AccountState:
    ExecutionState -> IExecutionState interface
    newContract field -> NewContract

As for tracer itself, StorageOpTracer handles storage operations. It uses MPTTracer to access the state. Thus, at the end of tracing it is able to get a trace for entire Contract MPT trie.