A proper workaround to the io_uring kernel bug
In #1298, we tried to work around the kernel bug with io_uring
that leads to the `npm ci` hanging in uninterruptible sleep.
That workaround has been inserted into the buildPhase, but
apparently `npm ci` is executed in the `unpackPhase`, so we
need to move the fix to the `preUnpackPhase` for it to have
the desired effect.