Fix and refactor on-chain Config
The issue was that the configuration was read by a non-main shard but written by the main shard.
However, the read operation was performed within an RW transaction, which caused a conflict in the Badger DB.
To solve this issue, we don't read config params through RW transaction anymore.
Now there are two Config accessors:
1. Read/write accessor. It is pretty much the same accessor that was previously, but it reads all config
   params to the internal map during initialization. This allows avoiding conflicts in the Badger.
2. Read accessor. This accessor gets RO transaction as input and reads params directly from MPT.
   It has much less overhead and can be used in read-only operations, like calling getter methods.

Also, creation of `ExecutionState` was refactored. Now there is only one function `NewExecutionState`
which accepts  `StateParams` struct to provide additional parameters for state creation.
It is convenient since there are a lot of cases in tests when most of the parameters can have default
values and can be just omitted.