[tests] make TestCliCometa more reliable
Sometimes it fails with an error:
```
Error: abi: attempting to unmarshall an empty string while arguments are expected
```

That means we try to call contract before it was included into main.
That's how eth_call works for now. This patch adds some delay to be sure
that contract is ready to use.