[execution] Fix big transactions processing
Previously, big transactions just stuck in the pool/blocks, and the node tried to execute them again and again.
Thus, a kind of infinite loop was observed.

This patch fixes it by adding constraint to the EVM that transaction cannot consume more than block gas limit.
Also, the check for exceeding the gas limit has now been moved after adding a transaction, otherwise such a
transaction would never appear in the block.

This solution has a drawback that the block can exceed its gas limit by two times in the worst case.
In the future, we'll probably want to solve it more appropriately by reducing the possible excess of the block
gas limit.

Relates to #79