[mpt] replace custom implementation with go-ethereum/trie
This patch removes our own MPT implementation in favour of well-known
and well-tested go-ethereum/trie module. Here we don't just replace it but
changes internal/mpt module as a wrapper of new implementation to avoid significant
changes in a codebase.

However still there are some differences in API, error handling that affected
that logic. MPT interface that is used as a storage wasn't significantly changed but
proof system was changed a bit. E.g. here we remove SimpleProof because initially we
worked with specific MPT nodes but now it's just abstract arrays of bytes.