drop external deployment
This patch removes external deployment from public interfaces (CLI, niljs)
and replaces it with deployment via faucet.
We need it to support EOA because each address with positive balance and without
contract code should be considered as EOA. But right now this address can be used
to deploy a contract.